fix: Fix the whole Sponsorship CRUD

+++ api/controllers/sponsorshipController.js
@@ -84,31 +84,28 @@ exports.read_a_sponsorship = async function (req, res) {
     if (err) {
       res.status(400).send(err);
     } else{
-      Sponsorship.find({_id: req.params.sponsorshipId, is_paid : true}, function (err, sponsorships) {
+
+      Sponsorship.findById(req.params.sponsorshipId, function (err, sponsorship) {
         if (err) {
-          res.status(400).send(err)
+          res.status(404).send(err)
         } else {
 
-          if(!sponsorships) {
-            res.status(404).send("The sponsorship does not exist or is not paid yet")
-          } else {
+          // check if sponsorship owner is me
+          if(sponsorship.actor == authSponsorId) {
 
-            if(sponsorships[0] != null) {
-              if(sponsorships[0].actor == authSponsorId) {
-                res.status(200).json(sponsorships[0])
-              }else{
-                res.status(400).send("The sponsorship does not exist or is not paid yet")
-              }
+            if(sponsorship.is_paid) {
+              res.status(200).json(sponsorship)
             } else {
-              res.status(400).send("The sponsorship does not exist or is not paid yet")
+              res.status(404).send("The sponsorship does not exist or is not paid yet")
             }
 
-            
+          } else {
+            res.status(400).send("This sponsor does not have permissions to update this sponsorship");
           }
 
-
         }
-      })
+      });
+
     }
   });
 
@@ -130,13 +127,22 @@ exports.update_a_sponsorship = async function (req, res) {
         if (err) {
           res.status(404).send(err)
         } else {
-          Sponsorship.findOneAndUpdate({ _id: req.params.sponsorshipId }, req.body, { new: true }, function (err, sponsorship) {
-            if (err) {
-              res.status(400).send(err)
-            } else {
-              res.status(201).json(sponsorship)
-            }
-          })
+
+          // check if sponsorship owner is me
+          if(sponsorship.actor == authSponsorId) {
+
+            Sponsorship.findOneAndUpdate({ _id: req.params.sponsorshipId }, req.body, { new: true }, function (err, sponsorship) {
+              if (err) {
+                res.status(400).send(err)
+              } else {
+                res.status(201).json(sponsorship)
+              }
+            })
+
+          } else {
+            res.status(400).send("This sponsor does not have permissions to update this sponsorship");
+          }
+
         }
       })
 
@@ -157,49 +163,69 @@ exports.delete_a_sponsorship = async function (req, res) {
       res.status(400).send(err);
     } else {
 
-      Sponsorship.deleteOne({
-        _id: req.params.sponsorshipId
-      }, function (err, sponsorship) {
+      Sponsorship.findById(req.params.sponsorshipId, function (err, sponsorship) {
         if (err) {
-          res.status(404).send(err)
+          res.status(404).send("Sponsor not found")
         } else {
 
-          // updated the sponsor sponsorships list
-          Actor.findOneAndUpdate({ _id: authSponsorId }, {"$pull": {sponsorships: req.params.sponsorshipId}}, { new: true }, function(err, result){
-            if(err){
-              res.send(err)
-            }
-            else{
-              res.status(204).json({ message: 'Sponsorship successfully deleted' })
-            }
-          })
-          
+          // check if sponsorship owner is me
+          if(sponsorship.actor == authSponsorId) {
+
+            Sponsorship.deleteOne({
+              _id: req.params.sponsorshipId
+            }, function (err, sponsorship) {
+              if (err) {
+                res.status(404).send(err)
+              } else {
+      
+                // updated the sponsor sponsorships list
+                Actor.findOneAndUpdate({ _id: authSponsorId }, {"$pull": {sponsorships: req.params.sponsorshipId}}, { new: true }, function(err, result){
+                  if(err){
+                    res.send(err)
+                  }
+                  else{
+                    res.status(204).json({ message: 'Sponsorship successfully deleted' })
+                  }
+                })
+                
+              }
+            })
+
+          }else {
+            res.status(400).send("This sponsor does not have permissions to delete this sponsorship");
+          }
+
         }
       })
+
 
     }
   });
 
 }
 
-// TODO: Implementar via PayPal
-exports.pay_a_sponsorship = function (req, res) {
-  Actor.findById(req.params.actorId, function (err, sponsor) {
+exports.pay_a_sponsorship = async function (req, res) {
+
+  // get auth sponsor
+  const idToken = req.header('idToken')
+  let authSponsorId = await authController.getUserId(idToken)
+  authSponsorId = String(authSponsorId)
+
+  Actor.findById(authSponsorId, function (err, sponsor) {
     if (err) {
       res.status(400).send(err);
     } else {
-      if (!sponsor) {
-        res.status(404).send("Sponsor not found");
-      } else if (!sponsor.role.find((role) => role === "SPONSOR")) {
-        res.status(400).send("Actor must be sponsor");
-      } else {
-        Sponsorship.findById(req.params.sponsorshipId, function (err, sponsorship) {
-          if (err) {
-            res.status(404).send("Sponsor not found")
+
+      Sponsorship.findById(req.params.sponsorshipId, function (err, sponsorship) {
+        if (err) {
+          res.status(404).send("Sponsor not found")
+        } else {
+          if(sponsorship.is_paid === true) {
+            res.status(400).send("Sponsorship is already payed");
           } else {
-            if(sponsorship.is_paid === true) {
-              res.status(400).send("Sponsorship is already payed");
-            } else {
+
+            // check if sponsorship owner is me
+            if(sponsorship.actor == authSponsorId) {
               Sponsorship.findOneAndUpdate({ _id: req.params.sponsorshipId }, {is_paid: true}, { new: true }, function (err, sponsorship) {
                 if (err) {
                   res.status(400).send(err)
@@ -207,10 +233,16 @@ exports.pay_a_sponsorship = function (req, res) {
                   res.status(201).json(sponsorship)
                 }
               })
+            } else {
+              res.status(400).send("This sponsor does not have permissions to pay this sponsorship");
             }
+
+            
           }
-        })
-      }
+        }
+      })
+
     }
   })
+
 }
